parsing the http request line
A little finnicky still, but it parses the http request line into
something that later can be used to reconstruct the original request or
pass through the request to another stream.

Eventually, the Wrapper should be responsible for creating a request to
a main/proxy server using that request line. The headers still need to
be inserted.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>