remove http clutter and clean up api
Remove the clutter that was in the http module and clean up the api to
have a better flow. The flow makes use of the builder pattern to prevent
early parsing of the http request. After the request is parsed, reuse
the request to send it to both main and shadow servers.

Fiddling with type state pattern vs builder pattern.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>