move to new http mod, use tryfrom for parsing
Move the parsing of http related stuff into it's own module and reexport
useful structs and functions.

Use TryFrom<String> to implement the parsing for all http related
structs. I don't know if this is what I want right now but lets see how
far we get with this version.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>