add default api of how the parsing should work
Add the api of how the parsing should work in the future. The
HttpRequest struct should be responsible for deciding whether the HTTP
request is done parsing. The  link provides a good base for parsing http
requests.

Send a 200 OK response to the client to test responsiveness of the
server. The response should obviously come from the main server, but for
now this is good enough to see program flow.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>