improve response passthrough, did not work properly
The response passthrough used read_buf, but that method is not nice for
speedy reading of the incoming bytes. For some reason, it hangs for a
few seconds waiting for more data, even though I expect servers to send
EOFs (I think Express does at least...).

It is fixed now by using the same reading loop as is done for the client
stream handling, except that it is still dirty copied from the client.
It should be more streamlined, and the resulting passthrough should be
checked as well.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>