add headers and content length to decoded http
The decoded http request should have content length readily available to
make it easier in the future when decoding the response part of the
response. The data is now treated exactly like a header, which should
not be the case.

Headers are decoded and inserted as a vector of header key value pairs.
The keys are an enum type to save space and (hopefully) decode faster.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>